mavlink: move COMMAND_LONG to separate stream header
